+++ database/migrations/2024_05_08_134946_create_order_details_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
                 $table->integer('order_id');
                 $table->integer('product_id');
                 $table->string('product_name');
-                $table->float('product_price');
+                $table->integer('product_price');
                 $table->integer('product_sales_quantily');
             $table->timestamps();
         });
